Fixing deadlock
Relaxing CountingSemaphore conditions to <=0 & >0 & changed the
initialization operation of the operators to have a ThreadPoolStage so
that There is no distributed deadlock involving the driver and the
restarted task. When the task restarts, its initialization loop is
sequential. So unless it receives TopoSetup for the first oper, it won't
proceed to the next. However, if the first operator has had its
TopoSetup on master it does not respond to further Additions/Deletions.
But the driver can proceed with other topologies if initialization is
multithreaded. However, this doesn't eliminate the possibility of a
data-flow deadlock where control operator has 0,1,2,4 as its children
but reduce has 0,1,2,3 then control msg was never sent to 3 but reduce
expects a response from 3. So we need to aggregate acknowledgement
processing so that all topologies are updated with the acknowledgements
in a guarded function and all topologies are in sync avoiding the
data-flow deadlock